backend/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
azure-core==1.26.1
azure-identity==1.11.0
azure-kusto-data==3.1.3
backcall==0.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
decorator==5.1.1
executing==1.2.0
flake8==5.0.4
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
frozenlist==1.3.3
greenlet==2.0.1
idna==3.4
ijson==3.1.4
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
msal==1.20.0
msal-extensions==1.0.0
multidict==6.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
portalocker==2.6.0
prompt-toolkit==3.0.32
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==2.6.0
pyserial==3.5
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.43
stack-data==0.6.0
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
yarl==1.8.1
